requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.7
colorama==0.4.6
DateTime==5.2
dotenv-python==0.0.1
exceptiongroup==1.1.3
fastapi==0.104.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
multidict==6.0.4
pamqp==3.2.1
pika==1.3.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
yarl==1.9.2
zope.interface==6.1
